remove temp debug logs
