+++ floors/floors.go
@@ -11,7 +11,6 @@ import (
 	"github.com/prebid/prebid-server/v3/metrics"
 	"github.com/prebid/prebid-server/v3/openrtb_ext"
 	"github.com/prebid/prebid-server/v3/util/ptrutil"
-	"github.com/prebid/prebid-server/v3/util/ulpdebug"
 )
 
 type Price struct {
@@ -293,18 +292,8 @@ func updateFloorsInRequest(bidRequestWrapper *openrtb_ext.RequestWrapper, priceF
 		}
 		prebidExt.Floors = priceFloors
 		requestExt.SetPrebid(prebidExt)
-		if ulpdebug.ShouldTrace(bidRequestWrapper.BidRequest) {
-			wiid := ulpdebug.Wiid(bidRequestWrapper.BidRequest)
-			ulpdebug.LogNote(wiid, "floors_update", "rebuild_request")
-		}
 		if err := bidRequestWrapper.RebuildRequest(); err != nil {
-			if ulpdebug.ShouldTrace(bidRequestWrapper.BidRequest) {
-				ulpdebug.LogStageErr(ulpdebug.Wiid(bidRequestWrapper.BidRequest), "floors_rebuild_request", err)
-			}
 			return
-		}
-		if ulpdebug.ShouldTrace(bidRequestWrapper.BidRequest) {
-			ulpdebug.LogStageOK(ulpdebug.Wiid(bidRequestWrapper.BidRequest), "floors_rebuild_request")
 		}
 	}
 }
