+++ util/rawext/rawext_test.go
@@ -1,4 +1,4 @@
-package ulpdebug
+package rawext
 
 import (
 	"encoding/json"
